loaded instacart data. Count number of aisles and arranged in descending order to see greatest to least. Made plot to show number of items order in each aisle. Need to arrange better sensibly.
Made table to show the three most popular items in the 3 specific aisles. However, I still cannot get the top three items. still needs to be edited.

Made second table showing the mean hour of the day at which Pink Lady Apples and Coffee Ice Cream are ordered on each day of the week. Renamed days of the week to be actual week names instead of numbers