made dataframe for states observed at 7 or more locations in 2002 and 2010.
Constructed dataset limited to excellent responses. contains year, state, and variable that averages data_value.

Made spaghetti plot. Still needs editting because values are vertical (instead of spaghetti style). To be updated.

Created dataframe for a two panel plot for the years 2006 and 2010 showing the distribution of data_value for Poor to Excellent responses.